requirements.txt
aiogram==2.25.1
motor
python-dotenv